fix(perm): fix file access

+++ NoAxiom/kernel/src/syscall/syscall.rs
@@ -21,7 +21,7 @@ impl<'a> Syscall<'a> {
         use SyscallID::*;
         match id {
             // fs
-            SYS_FCHMODAT =>         self.sys_fchmodat(args[0], args[1], args[2], args[3]),
+            SYS_FCHMODAT =>         self.sys_fchmodat(args[0], args[1], args[2], args[3] as i32),
             SYS_FCHOWNAT =>         Self::empty_syscall("fchownat", 0),
             SYS_UMASK =>            Self::empty_syscall("umask", 0x777),
             SYS_SYNC =>             Self::empty_syscall("sync", 0),
